synthetic_data_resume_coach/pyproject.toml
[tool.uv]
python = "3.12"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "synthetic-data-resume-coach"
version = "0.1.0"
description = "Synthetic data pipeline for generating and validating resume + job description pairs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic[email]>=2.0",
    "openai>=1.0",
    "llm-utils",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "braintrust>=0.1",
    "logfire>=0.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "pre-commit>=3.7",
    "black>=24.4",
    "ruff>=0.4",
    "mypy>=1.10",
    "types-requests>=2.31",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["venv", ".venv"]
